requirements.txt
Django>=4.2,<4.3
pillow>=10.4,<10.5
psycopg2>=2.9,<2.10
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
django-phonenumber-field[phonenumberslite]>=7.3,<7.4
python-slugify>=8.0,<8.1
django-debug-toolbar>=4.3,<4.4
drf-spectacular>=0.27,<0.28
coverage>=7.6,<7.7
django-filter>=23.5,<24.6
django-redis>=5.4,<5.5
stripe>=8.4,<8.5
sendgrid>=6.11,<6.12





